refactor: Extract modules to files
