+++ iox_objectstore_garbage_collect/src/main.rs
@@ -27,6 +27,10 @@ use snafu::prelude::*;
 use std::{process::ExitCode, sync::Arc};
 use tokio::sync::mpsc;
 
+mod checker;
+mod deleter;
+mod lister;
+
 const BATCH_SIZE: usize = 1000;
 
 fn main() -> ExitCode {
@@ -121,173 +125,3 @@ enum Error {
 }
 
 type Result<T, E = Error> = std::result::Result<T, E>;
-
-mod lister {
-    use futures::prelude::*;
-    use object_store::ObjectMeta;
-    use snafu::prelude::*;
-    use std::sync::Arc;
-    use tokio::sync::mpsc;
-
-    pub(crate) async fn perform(
-        args: Arc<crate::Args>,
-        checker: mpsc::Sender<ObjectMeta>,
-    ) -> Result<()> {
-        let object_store = args
-            .object_store()
-            .await
-            .context(CreatingObjectStoreSnafu)?;
-
-        let mut items = object_store.list(None).await.context(ListingSnafu)?;
-
-        while let Some(item) = items.next().await {
-            let item = item.context(MalformedSnafu)?;
-            checker.send(item).await?;
-        }
-
-        Ok(())
-    }
-
-    #[derive(Debug, Snafu)]
-    pub(crate) enum Error {
-        #[snafu(display("Could not create the object store"))]
-        CreatingObjectStore {
-            source: clap_blocks::object_store::ParseError,
-        },
-
-        #[snafu(display("The prefix could not be listed"))]
-        Listing { source: object_store::Error },
-
-        #[snafu(display("The object could not be listed"))]
-        Malformed { source: object_store::Error },
-
-        #[snafu(display("The checker task exited unexpectedly"))]
-        #[snafu(context(false))]
-        CheckerExited {
-            source: tokio::sync::mpsc::error::SendError<ObjectMeta>,
-        },
-    }
-
-    pub(crate) type Result<T, E = Error> = std::result::Result<T, E>;
-}
-
-mod checker {
-    use object_store::ObjectMeta;
-    use snafu::prelude::*;
-    use std::sync::Arc;
-    use tokio::sync::mpsc;
-
-    pub(crate) async fn perform(
-        args: Arc<crate::Args>,
-        mut items: mpsc::Receiver<ObjectMeta>,
-        deleter: mpsc::Sender<ObjectMeta>,
-    ) -> Result<()> {
-        let catalog = args.catalog().await.context(CreatingCatalogSnafu)?;
-        let cutoff = args.cutoff();
-
-        let mut repositories = catalog.repositories().await;
-        let parquet_files = repositories.parquet_files();
-
-        while let Some(item) = items.recv().await {
-            if item.last_modified < cutoff {
-                // Not old enough; do not delete
-                continue;
-            }
-
-            let file_name = item.location.parts().last().context(FileNameMissingSnafu)?;
-            let file_name = file_name.to_string(); // TODO: Hmmmmmm; can we avoid allocation?
-
-            if let Some(uuid) = file_name.strip_suffix(".parquet") {
-                let object_store_id = uuid.parse().context(MalformedIdSnafu { uuid })?;
-                let parquet_file = parquet_files
-                    .get_by_object_store_id(object_store_id)
-                    .await
-                    .context(GetFileSnafu { object_store_id })?;
-
-                if parquet_file.is_some() {
-                    // We have a reference to this file; do not delete
-                    continue;
-                }
-            }
-
-            deleter.send(item).await.context(DeleterExitedSnafu)?;
-        }
-
-        Ok(())
-    }
-
-    #[derive(Debug, Snafu)]
-    pub(crate) enum Error {
-        #[snafu(display("Could not create the catalog"))]
-        CreatingCatalog {
-            source: clap_blocks::catalog_dsn::Error,
-        },
-
-        #[snafu(display("Expected a file name"))]
-        FileNameMissing,
-
-        #[snafu(display(r#""{uuid}" is not a valid ID"#))]
-        MalformedId { source: uuid::Error, uuid: String },
-
-        #[snafu(display("The catalog could not be queried for {object_store_id}"))]
-        GetFile {
-            source: iox_catalog::interface::Error,
-            object_store_id: uuid::Uuid,
-        },
-
-        #[snafu(display("The deleter task exited unexpectedly"))]
-        DeleterExited {
-            source: tokio::sync::mpsc::error::SendError<ObjectMeta>,
-        },
-    }
-
-    pub(crate) type Result<T, E = Error> = std::result::Result<T, E>;
-}
-
-mod deleter {
-    use object_store::ObjectMeta;
-    use snafu::prelude::*;
-    use std::sync::Arc;
-    use tokio::sync::mpsc;
-
-    pub(crate) async fn perform(
-        args: Arc<crate::Args>,
-        mut items: mpsc::Receiver<ObjectMeta>,
-    ) -> Result<()> {
-        let object_store = args
-            .object_store()
-            .await
-            .context(CreatingObjectStoreSnafu)?;
-        let dry_run = args.dry_run;
-
-        while let Some(item) = items.recv().await {
-            let path = item.location;
-            if dry_run {
-                eprintln!("Not deleting {path} due to dry run");
-            } else {
-                object_store
-                    .delete(&path)
-                    .await
-                    .context(DeletingSnafu { path })?;
-            }
-        }
-
-        Ok(())
-    }
-
-    #[derive(Debug, Snafu)]
-    pub(crate) enum Error {
-        #[snafu(display("Could not create the object store"))]
-        CreatingObjectStore {
-            source: clap_blocks::object_store::ParseError,
-        },
-
-        #[snafu(display("{path} could not be deleted"))]
-        Deleting {
-            source: object_store::Error,
-            path: object_store::path::Path,
-        },
-    }
-
-    pub(crate) type Result<T, E = Error> = std::result::Result<T, E>;
-}
